fix(gateway): keep session method metadata lightweight
Separate unchanged static session target policy from runtime target resolution so scope and profile metadata no longer import session storage. Preserve authorization and legacy profile-independent methods, and guard the metadata import boundary. Measured cold skill-test startup improves without changing timeouts, mocks, or setup; historical timeout attribution remains uncertain under host contention.
